Move artifact-related code to internal
