+++ internal/artifact/artifact_uploader_test.go
@@ -1,4 +1,4 @@
-package agent
+package artifact
 
 import (
 	"context"
