+++ internal/artifact/s3_uploader.go
@@ -1,4 +1,4 @@
-package agent
+package artifact
 
 import (
 	"context"
